Resolve jerky scroll behaviour on window focus change
A strange jerky behaviour occured when scrolling down, unfoccusing the
window and then on refocussing scrolling up again where the scroll would
start from the initial position before the initial scroll down. This
seemed to be related to the previous state of the animationFrame not
being completed. By setting it to null this bug seems to be resolved.