Implement UI in SwiftUI
Transitions the current UI, using Cocao and the storyboard in to using
SwiftUI. The UI is still loaded in to a storyboard using the
NSHostingController. It would likely be possible to remove the need for
this storyboard as well.

My main reason for doing this is that I'm familiar with SwiftUI. By
making this change I will be able to further customise the interface

This change means that the minimum supported version of Vimari goes up
to OSX 10.15.0+. Therefore I don't expect this change to be merged soon.